Refactor: Refactor pycjs kernel.
